+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Stack `tool_version` in `terraplane.yaml`
 - `SCM_PROVIDER` ENV var to configure the chosen SCM provider at runtime
 
 ### Changed